requirements.txt
aws-cdk-lib==2.234.1
constructs>=10.0.0,<11.0.0
cdk_nag
